+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Exclamation symbols (:exclamation:) note something of importance e.g. breaking c
 ## [Unreleased]
 ### Added
 - Code checkers to ensure coding standard.
+- When releasing a new version of the Support Bot, automatically fetch the latest code and install with composer.
 ### Changed
 - Bumped Manager to 1.5.
 - Logging is now decoupled with custom Monolog logger.